# services/web/server/src/simcore_service_webserver/api/v0/openapi.yaml
openapi: 3.0.2
info:
  title: simcore-service-webserver
  description: Main service with an interface (http-API & websockets) to the web front-end
  version: 0.38.0
servers:
- url: ''
  description: webserver
- url: http://{host}:{port}
  description: development server
  variables:
    host:
      default: localhost
    port:
      default: '8001'
paths:
...
  /v0/catalog/services/{service_key}/{service_version}:
    get:
      tags:
      - catalog
      summary: Get Service
      operationId: get_service
      parameters:
      - required: true
        schema:
          title: Service Key
          pattern: ^simcore/services/((comp|dynamic|frontend))/([a-z0-9][a-z0-9_.-]*/)*([a-z0-9-_]+[a-z0-9])$
          type: string
        name: service_key
        in: path
      - required: true
        schema:
          title: Service Version
          pattern: ^(0|[1-9]\d*)(\.(0|[1-9]\d*)){2}(-(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*)(\.(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*))*)?(\+[-\da-zA-Z]+(\.[-\da-zA-Z-]+)*)?$
          type: string
        name: service_version
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ServiceGet_'
    patch:
      tags:
      - catalog
      summary: Update Service
      operationId: update_service
      parameters:
      - required: true
        schema:
          title: Service Key
          pattern: ^simcore/services/((comp|dynamic|frontend))/([a-z0-9][a-z0-9_.-]*/)*([a-z0-9-_]+[a-z0-9])$
          type: string
        name: service_key
        in: path
      - required: true
        schema:
          title: Service Version
          pattern: ^(0|[1-9]\d*)(\.(0|[1-9]\d*)){2}(-(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*)(\.(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*))*)?(\+[-\da-zA-Z]+(\.[-\da-zA-Z-]+)*)?$
          type: string
        name: service_version
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ServiceGet_'
...
  /v0/catalog/services/{service_key}/{service_version}/pricing-plan:
    get:
      tags:
      - catalog
      - pricing-plans
      summary: Retrieve default pricing plan for provided service
      operationId: get_service_pricing_plan
      parameters:
      - required: true
        schema:
          title: Service Key
          pattern: ^simcore/services/((comp|dynamic|frontend))/([a-z0-9][a-z0-9_.-]*/)*([a-z0-9-_]+[a-z0-9])$
          type: string
        name: service_key
        in: path
      - required: true
        schema:
          title: Service Version
          pattern: ^(0|[1-9]\d*)(\.(0|[1-9]\d*)){2}(-(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*)(\.(0|[1-9]\d*|\d*[-a-zA-Z][-\da-zA-Z]*))*)?(\+[-\da-zA-Z]+(\.[-\da-zA-Z-]+)*)?$
          type: string
        name: service_version
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ServicePricingPlanGet_'
...
  /v0/computations/{project_id}:start:
    post:
      tags:
      - computations
      - projects
      summary: Start Computation
      operationId: start_computation
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComputationStart'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope__ComputationStarted_'
  /v0/computations/{project_id}:stop:
    post:
      tags:
      - computations
      - projects
      summary: Stop Computation
      operationId: stop_computation
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '204':
          description: Successful Response
...
  /v0/projects/{project_id}:xport:
    post:
      tags:
      - projects
      - exporter
      summary: Export Project
      description: creates an archive of the project and downloads it
      operationId: export_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
...
  /v0/projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: list_projects
      parameters:
      - description: Order by field (type|uuid|name|description|prj_owner|creation_date|last_change_date)
          and direction (asc|desc). The default sorting order is ascending.
        required: false
        schema:
          title: Order By
          description: Order by field (type|uuid|name|description|prj_owner|creation_date|last_change_date)
            and direction (asc|desc). The default sorting order is ascending.
          default: '{"field": "last_change_date", "direction": "desc"}'
        example: '{"field": "last_change_date", "direction": "desc"}'
        name: order_by
        in: query
      - required: false
        schema:
          title: Limit
          exclusiveMaximum: true
          minimum: 1
          type: integer
          default: 20
          maximum: 50
        name: limit
        in: query
      - required: false
        schema:
          title: Offset
          minimum: 0
          type: integer
          default: 0
        name: offset
        in: query
      - required: false
        schema:
          allOf:
          - $ref: '#/components/schemas/ProjectTypeAPI'
          default: all
        name: type
        in: query
      - required: false
        schema:
          title: Show Hidden
          type: boolean
          default: false
        name: show_hidden
        in: query
      - required: false
        schema:
          title: Search
          maxLength: 100
          type: string
        name: search
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ProjectListItem_'
    post:
      tags:
      - projects
      summary: Creates a new project or copies an existing one
      operationId: create_project
      parameters:
      - required: false
        schema:
          title: From Study
          type: string
          format: uuid
        name: from_study
        in: query
      - required: false
        schema:
          title: As Template
          type: boolean
          default: false
        name: as_template
        in: query
      - required: false
        schema:
          title: Copy Data
          type: boolean
          default: true
        name: copy_data
        in: query
      - required: false
        schema:
          title: Hidden
          type: boolean
          default: false
        name: hidden
        in: query
      requestBody:
        content:
          application/json:
            schema:
              title: ' Create'
              anyOf:
              - $ref: '#/components/schemas/ProjectCreateNew'
              - $ref: '#/components/schemas/ProjectCopyOverride'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TaskGet_'
...
  /v0/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get Project
      operationId: get_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
    put:
      tags:
      - projects
      summary: Replace Project
      description: Replaces (i.e. full update) a project resource
      operationId: replace_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectReplace'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '204':
          description: Successful Response
    patch:
      tags:
      - projects
      summary: Update Project
      description: Partial update of a project resource
      operationId: update_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectGet_'
  /v0/projects/{project_id}:clone:
    post:
      tags:
      - projects
      summary: Clone Project
      operationId: clone_project
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TaskGet_'
...
  /v0/projects/{project_id}/metadata:
    get:
      tags:
      - projects
      - metadata
      summary: Get Project Metadata
      operationId: get_project_metadata
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectMetadataGet_'
    patch:
      tags:
      - projects
      - metadata
      summary: Update Project Metadata
      operationId: update_project_metadata
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMetadataUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ProjectMetadataGet_'
...
  /v0/projects/{project_id}/nodes/{node_id}/outputs:
    patch:
      tags:
      - projects
      - nodes
      summary: Update Node Outputs
      operationId: update_node_outputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
        name: project_id
        in: path
      - required: true
        schema:
          title: Node Id
          type: string
        name: node_id
        in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeOutputs'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
...
  /v0/projects/{project_id}/nodes/{node_id}/pricing-unit:
    get:
      tags:
      - projects
      summary: Get currently connected pricing unit to the project node.
      operationId: get_project_node_pricing_unit
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      - required: true
        schema:
          title: Node Id
          type: string
          format: uuid
        name: node_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Union_PricingUnitGet__NoneType__'
...
  /v0/projects/{project_id}/outputs:
    get:
      tags:
      - projects
      - ports
      summary: Get Project Outputs
      description: New in version *0.10*
      operationId: get_project_outputs
      parameters:
      - required: true
        schema:
          title: Project Id
          type: string
          format: uuid
        name: project_id
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_dict_uuid.UUID__simcore_service_webserver.projects._ports_handlers.ProjectOutputGet__'
...
components:
  schemas:
...
    ComputationStart:
      title: ComputationStart
      type: object
      properties:
        force_restart:
          title: Force Restart
          type: boolean
          default: false
        cluster_id:
          title: Cluster Id
          minimum: 0
          type: integer
          default: 0
        subgraph:
          title: Subgraph
          uniqueItems: true
          type: array
          items:
            type: string
          default: []
...
    Envelope_ServiceGet_:
      title: Envelope[ServiceGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/models_library__api_schemas_webserver__catalog__ServiceGet'
        error:
          title: Error
...
    Envelope_ServicePricingPlanGet_:
      title: Envelope[ServicePricingPlanGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ServicePricingPlanGet'
        error:
          title: Error
...
    Envelope_TaskGet_:
      title: Envelope[TaskGet]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskGet'
        error:
          title: Error
...
    Envelope__ComputationStarted_:
      title: Envelope[_ComputationStarted]
      type: object
      properties:
        data:
          $ref: '#/components/schemas/_ComputationStarted'
        error:
          title: Error
...
    ProjectTypeAPI:
      title: ProjectTypeAPI
      enum:
      - all
      - template
      - user
      type: string
      description: An enumeration.
...